url shortener. After review refactoring HashCache interface added, Cache check and refill invoke logic changed scheduling settings refactored
